test: divide integration test files according to different cases
skeleton:

1. test_draft

    - test_commit_draft
    - test_create_draft
    - test_list_drafts

2. test_tag

    - test_create_tag
    - test_delete_tag
    - test_get_tag
    - test_list_tags

3. test_commit

    - test_get_commit
    - test_get_commit_by_ref
    - test_list_commits
    - test_draft_and_top_commit_inherit

4. test_branch

    - test_get_branch
    - test_list_branches

5. test_data

    - test_delete_data
    - test_list_data_paths
    - test_list_file_order
    - test_replace_file
    - test_replace_label

6. test_sensor

    - test_sensor

7. test_segment

    - test_create_segment
    - test_delete_segment
    - test_get_or_create_segment
    - test_get_or_create_fusion_segment
    - test_list_segment_names

8. test_dataset

    - test_create_dataset
    - test_create_dataset_with_region
    - test_create_fusion_dataset
    - test_delete_dataset
    - test_get_dataset_to_latest_commit
    - test_get_fusion_dataset_to_latest_commit
    - test_get_new_dataset
    - test_list_dataset_names
    - test_rename_dataset
    - test_catalog
    - test_checkout
    - test_notes

9. test_upload_dataset

    - test_upload_dataset_only_with_file
    - test_upload_dataset_to_given_draft
    - test_upload_dataset_with_label

10. test_upload_segment

    - test_get_and_upload_fusion_segment_only_with_file
    - test_get_and_upload_fusion_segment_with_label
    - test_get_and_upload_fusion_segment_with_no_file
    - test_get_and_upload_segment_only_with_file
    - test_get_and_upload_segment_with_label
    - test_get_and_upload_segment_with_no_file

11. test_upload_data

    - test_add_file
    - test_upload_data_with_label
    - test_upload_data_without_label
    - test_upload_file
    - test_upload_label
    - test_upload_label_without_catalog

12. test_upload_frame
    - test_upload_frame_with_label
    - test_upload_frame_with_order
    - test_upload_frame_without_label
    - test_upload_frame_without_sensor

PR Closed: https://github.com/Graviti-AI/tensorbay-python-sdk/pull/535
